More extracting of inner classes to outer classes for package cleanup.
